🔧(styles) enable purgeCSS to reduce tailwind build size
By default tailwind generates very large CSS bundle sizes as it
includes a large quantity of utility classes, 99% of which are not
actually used in Partaj. This is great in development but not what
we want in production.

There, we want to enable purgeCSS to remove all unused classes. We
added our own way to determine whether to enable it or not using a
flag, rather than do gymnastics with our Dockerfile for NODE_ENV.