🐛(timezones) force timezone on dates in referral lists
We setup Django to use timezone aware datetimes. This means all our
datetimes are storing the correct information. This is reflected in
views which render dates in the frontend: the browser's timezone info
is used and the datetime is correct, timezone and all.

As for backend templates however, we need to implement custom logic to
handle timezones using sessions or whatever, but that still requires
using some bit of JS to pick up the browser timezone in the first place.

We don't want to implement this as we project the only two backend-rendered
views that currently display datetimes (referral lists for requester and
unit) will be migrated to the frontend at some point, which handles this
seamlessly.

Therefore, we decided to just force the timezone in those two views to
"fix" the bug for now, instead of diving in an implementation that should
only be temporary anyway.