✨(email) update the "referral answered" email template
When we created the "referral answered" email template, we included
the referral answer content directly into the body of the email. We
knew it was not a good solution, but we did not have any other view
to put referral answers.

Now that we have working views to see referrals on the requester
side, it is high time we replaced the answer content with a CTA in
the "referral answer" email template.