💄(frontend) improve styles for the list of attached files
When we refactored Partaj to stop using Bootstrap, we forgot to
rework the styles for the list of attachments to a Referral. This
is correcting that mistake and improving on `list-group-item` to
make them more usable & accessible.

We also took this opportunity to add an outline to the "Answer"
button at the bottom of the referral as we noticed it was missing
it.