🔧(environment) add distributable env files to the repo
In development with docker-compose, we use environment files to
pass information through to our Django backend. The env.d directory
needs to be gitignored so we avoid committing secrets by mistake.

However, we still need to provide the test env file as well as
an example of what the development environment file should look
like.