📝(docs) add a README.md with instructions on getting started
As it stood, the project was complicated to get running as it was
missing a README file and any kind of documentation. Adding a
README with information on architecture, how to get started, and
how to perform development is a first step in making it usable
by anyone beyond the initial contributor.