♻️(settings) rework emailing related settings
We determined our old approach of passing template IDs through
environment variables was not appropriate: as the use of so or so
template version is compatible with a given version of the code,
those template numbers should be included in the code itself.

We therefore removed some expected environment variables and moved
them to a settings mixin in which they are hardcoded. This keeps a
clear separation with the actual emailer implementation while not
artificially delegating the responsibility to the environment.

We also added the two relevant environment variables (provider API
url and API key) to this new mixin, though they are still picked up
from env variables.