✅(frontend) add tests for <ReferralDetailContent />
As we built <ReferralDetailContent /> during a sprint to implement
more interaction through referral assignments/unassignments on the
referral detail page, when we needed to rework the whole referral
detail view to build it in React, it was not tested as we implemented
it.

We took some time to add a simple unit test for it, and found errors
that we took this chance to fix.