✅(frontend) add tests for <ReferralActivityDisplay />
We recently added new strings to display activities for
assignments & unassignments where the actor is the same as the
user being assigned/unassigned.

As we did, we noticed the <ReferralActivityDisplay /> component
was not tested at all, even though it's at the center of Partaj.
We added some tests for all the activities we're currently supporting.