✅(frontend) add a yarn script to collect coverage info
Coverage can sometimes be useful to collect to check if some code
paths have been inadvertantly forgotten in tests. It can't be the
default however as it slows down tests.

We're adding a yarn script command to run tests with coverage &
gitignoring the results so as not to commit them by mistake.