🐛(users) fix migration 0002_createsuperuser
Using a migration to create the initial superuser is a practical way
to always have an admin account when starting a new server. It's
particularly useful with PaaS instances on which it may not be
trivial or even possible to access the instance to run python commands.

However, our own migration suffered from 2 issues, which became
apparent after we made changes to the User model:
- it imported the User model through `get_user_model` which gets the
  "current" user model as in, "current" in the codebase. This is not
  appropriate for data migrations as they need to have models that
  reflect the state of the schema at the time of the migration;
- it used the `createsuperuser` method, which works fine with the
  "current" User model from the code, but is not available in the faked
  past versions of models availables in migrations (see #26445).

We updated the migration to use the "fake past version" of the user
model and bypass `createsuperuser`.