🗃(migrations) make user & topic non-nullable on referral
Frontend and backend code in Partaj operate on the assumption that
the user & topic fields are present. Allowing them to be null only
invites problems because it creates a disrepancy between database
constraints and code expectations.

As all referrals in our environments have values for user & for
topic, we can just make these fields required without further changes.