♻️(forms) add user to our ReferralForm
We recently changed the Referral model to make user and topic
required fields. For this to work properly with our referral creation
method (through a ModelForm), we need to add the current user to
the ReferralForm instead of patching it in after creating the
Referral.

It's also cleaner as it creates a full referral from the start instead
of creating a temporary incorrect state.