🗃(tables) update table names for referral activities & answers
All our table names are prefixed with `partaj_`. When we created the
answer and the activity models, we forgot to add a Meta to name their
tables and add the prefix.

Remedy this mistake with a migration. We also took this opportunity to
add string representation for both answers and activities.

NB: we also had to fix the way models are imported in a previous
migration that was broken with the current models.