♻️(models) rework ReferralAttachment to use an abstract base class
We need to support attachments to other kinds of objects than just
Referral. The first one is ReferralAnswer, although there may be
more later on.

Unlike Activity items, we don't foresee a lot of different arbitrary
objects we'd need to keep track on on attachments. Therefore, we can
use an abstract base class for Attachment and create models based on
it depending on what kind of object the attachment is linked to.

This also allows us to do this refactor transparently (apart from some
docstrings on ReferralAttachment fields that mention referral).