✨(models) add a referral answer attachment model
In addition to attachments to referral, we now need to manage
attachments to referral answers. We're using the abstract base class
we build from our initial ReferralAttachment implementation so we
only have to manage the "referral_answer" field on the specific
model.

We also had to update the attachment view to be more generic in
the way it looks for an attachment matching the UUID in the url.