🍱(icons) add a "trash" icon
As we use a react-dropzone powered custom component to handle referral
answer attachments, instead of the native widget we used in the
referral form itself, we need to give users a way to see and remove
any file they already added. This is where we need this new "trash"
icon.