➕(frontend) add react-dropzone & filesize to our dependencies
To power our frontend referral answer form, we need a drag-and-drop
component that also lets users easily open the native widget to pick
files.

We chose react-dropzone as it provides hooks to easily implement
this and allows us to build it however we want. We also added filesize
as we want to be sure to display file sizes for the selected attachments
in a human-readable way.