✨(backend) allow our "answer" action to receive attachments
As we have a ReferralAnswerAttachment model that works just like our
established ReferralAttachment model, we simply have to update the
answer methods on the referral model and viewset to accept files
and create the relevant model instances for every attachment.

Django-rest-framework should take care of managing multipart form
input instead of plain JSON for us.