✨(api) open read endpoints for topics
As we refactor the referral creation form to be frontend-based
instead of a Django-form (allowing us to make a better upload UI, and
most importantly to add Rich-Text editing), we need to list the
available topics to show them in a frontend component.

We're only opening read endpoints no as we do not need to change
topics from the API right now.