🐛(files) fix an issue with some non-latin-1 filenames
We used a manually built Content-Disposition header to hint browsers
to treat attachments as files and download them. We missed that the
HTTP headers only support latin-1 characters, which do not include eg.
french accented letters.

To support this, we need to use "filename=*". It turns out we can
delegate the setting of this header to Django by using options to
FileResponse instead of building it ourselves.