♿️(frontend) add suport for aria-describedby to our attachments field
We built our <AttachmentsFormField /> with support for "aria-labelledby",
which we need to give a name to our attachments field. In the referral
form, we're also using descriptive "help text" along with our fields. This
kind of help can be linked to the field through "aria-describedby". We're
adding optional support for it to the component.