✨(api) add an endpoint to get a list of possible urgencies
As of this commit, possible values for the urgency of a referral
are hardcoded as part of the Referral model. We intend to move them
to a dedicated model that will allow users to manage them more easily.

To make that transition easier, we're opening an endpoint to get the
static values into the frontend, instead of copy-pasting them. This way,
when we change them or more them to a model, the frontend should remain
entirely unaffected.