➕(backend) add dj-inmemorystorage to our dependencies
We're adding tests for Referral creation now that it is handled
by an API call. As part of these tests, we want to make sure file
upload works as intended, which is difficult to achieve as files
are stored on s3-like remote services.

`dj-inmemorystorage` enables us to replace the Boto3 file storage
with in-memory storage so we can run our tests "as if" uploaded
files were available.