✨(api) add a CREATE handler for Referrals on the API
As we refactor our Django view for referral creation into a react
driven frontend view, we need to move the code that handles it
from a Django view to a DRF API method. For the most part, we can
duplicate the code from the Django view (which will be removed after
we complete our refactor).

We also took this opportunity to add tests for our new API endpoint,
which were missing on the Django view.