✨(frontend) build out a rich text editor with PM's example-setup
In an ideal world, we'd using an off-the-shelf solution to build
our editor. None of the ones we found (with acceptable licences for
our use case) were appropriate. This is why we chose ProseMirror.

While a promising React-based library based on PM (remirror) is
under development, it is not ready for us at this time, which is
why we went with the plain Javascript implementation. Going forward,
we plan on switching to remirror as it gets ready for production.

Otherwise, we can always improve this plain Javascript basis to
fit our needs going forward.