✨(frontend) update <ReferralForm /> to use rich text fields
The more long-form fields on Referral objects are appropriate places
to use a Rich text field: they can definitely benefit from some
structure. However, as they are already nested and broken down into
the structure of the Referral itself, in our opinion they do not need
actual heading elements.

We're therefore using a more bare-bones schema and menu to enable them
while not breaking page structure for referrals.

None of this requires any changes to the backend as those were already
TextFields, meaning strings with unlimited length. They can just as well
contain serialized JSON (which the backend is not expected to perform
any action on or query upon, so we don't need JSON fields either).

The only improvement we might be missing here is to enable the django
admin to edit rich text with WYSIWYG. However, that would be too much
duplicated effort for something we do not need right now.