✨(frontend) add a component to display ProseMirror rich text
As we're able to let users write rich text through ProseMirror and
serialize it to save it on referrals, we need to be able to parse
that and display it on the other side.

Unfortunately, ProseMirror does not come with "view" factilities out
of the box, which is why we built our own parsing logic matching our
current use-case and schema. Fortunately, the serialization schema
used by PM is simple enough to iterate through to generate the expected
HTML through react components.