🐛(frontend) fix missing element on list queries
We have not implemented any kind of pagination on our list requests
yet (on the frontend side). This was causing an issue with some of
them, eg. the topics list on a new referral, where it was only getting
and displaying the first 10 topics.

Set a limit to 999 by default so we can ignore this issue for now and
handle pagination/scaling properly later on, on a case by case basis.