✨(frontend) improve rich-text styles
To build our rich text editor, we relied on defaults for code and
styles from ProseMirror. As that does not include any styling or
feature to display the rich text generated through PM, we need to
add our own styles to the generated markup.

We're also applying it to the content inside the PM editor to
ensure what users see is what they really get.