✨(frontend) make referral answer use a rich text field
Referral answers, as another other place with longer form
user-entered text beyond referral fields, also need rich text to
enable users to give more structure to the answers.

As answers are not structured through a bunch of categories,
we enabled heading levels to allow users to structure the answers
themselves.