🌐(frontend) enable localization of PM based editor
Our PM based editor does not "live in React world" and uses some
defaults picked up from ProseMirror packages to get started more
easily. This means it is lacking translation facilities that we
use elsewhere in Partaj.

We can patch this by passing an "intl" provider into our setup call
and using it to get messages for our own editor UI elements, and monkey
patch translated messages onto the ones we took from the PM packages
as-is.

Later on, if we end up keeping this non-react implementation of PM
(as opposed to switching to remirror when it becomes production ready),
we'll have to refactor it into something more robust. It should
however be enough for our first implementation.