🐛(editor) use history info doc state polling
As we implemented our ProseMirror editor, we did not find a hook
to be informed (and pass it back to our React components) when the
editor's internal state is changed.

We resorted to polling for changes in the state. As a first
approximation, we used content length to try to detect if the user
interacted with the editor. This approximation turned out to be lossy
as it fails to detect styling changes.

Instead, we can rely on the history last change timestamp, which,
while it is admittedly a hack, relies on a more reliable source to
detect user changes.