✨(api) update the ReferralUrgency endpoint to return model instances
As we built a model (ReferralUrgency) to handle urgencies from the
back-office and better manage them in general, instead of relying
on hardcoded values on the model, we need to make sure the API
endpoint for referral urgencies is able to return them.

We did not use django-parler-rest as it returns a dictionary of
available translations. We want to return just the translation that
is relevant for the current request.