✨(frontend) display model based ReferralUrgency on referral detail
Update the API to return a serialized ReferralUrgency instance on
Referral objects on the API. Then, we can update the Referral type
in the frontend so it's able to actually display referrals with
their model based urgency levels.

We also updated the frontend factory for referral urgency so we can
keep our tests up to date.