✨(form) force justification for urgency levels that require it
We built a ReferralUrgency model to help us categorize our urgency
levels between the default one and the ones that require a
justification.

Now that the frontend has this information, we were able to update
the <ReferralForm /> to prevent submission until there is a
justification, if the selected urgency level requires it.