✨(referral) add an action endpoint to publish an answer
As we now only create draft answers when a unit member creates
an answer, and we do not intend to show these drafts to non members
of the referral answering unit, we need to add a method to publish
a given answer and make it available for the requester.

We also built a ReferralAnswerFactory to help us test this new
API route and model method.