🗃(models) add a recursive link on answers for draft<=>published
We want our published answers to have a relationship that allows us
to directly get the draft version, and vice versa. One example of a
feature that creates that need is that validations will be linked
to the draft version, but we'll often want to show them along with
the published version. We don't want to have to duplicate them also.

To this end, we created a recursive relationship that will be used to
represent the link between a draft & a published answer.