🗃(referral) update models for referral answer validations
Referral answer validations can almost be supported by adding
new activity types. Those provide us with the timing of the
validation request & validation itself, and can also link the
requester with the request and the validation with the validator.

The only missing piece is that we need the request to link the
validator and answer together. To that end, we're adding a new,
light model that only exists to represent that link by using
two foreign keys.