✨(frontend) show error messages when required fields are missing
The referral form errors out without any warning when a required
field using rich text is not filled. In this case, we want to show
an error message instead.

We updated our machines to issue a general "clean" signal to all
fields so they show their respective error messages when that
happens. We also added custom error messages to the rich text
fields.