✨(api) add an endpoint to request answer validations
As we build out our workflow to manage answer validations, we need
an endpoint for users to request a new validation, which will be
used for the first part of a validation.

We added a self-transition that locks this action to the appropriate
state on the referral, a matching API endpoint with error handling and
tests to make sure all our bases are covered.