✨(api) open readonly endpoints for referral activities
We need to remove referral activities from the referral objects as
we return them on the API: different users are allowed to see
different sets of activities and this creates too much complexity
to manage it as embedded objects on a serializer.

Instead, we can let users make LIST requests on an activities
endpoint and then use our logic to determine which activities they
are allowed to see.