✅(factories) add a factory for ReferralActivity
As we wrote our tests for the new referral activity API endpoints,
we needed a factory to be able to quickly create activities.
We added a post generation hook to make our fake activities more
realistic by adding the relevant object on the generic foreign key
depending on the type of activity (ie. its verb).