🔥(api) remove referral activities from referral objects
As we added a new endpoints we intend to use to get the activity
for a referral, instead of using the activities embedded on a
referral object, we can remove them here.

None of the underlying relationships are changed, we are only
changed what is returned on the API here.