✨(frontend) create data hooks using react-query
We have started using react-query in some of our views. However, if
we want to get the full benefit of using it as we move forward and
hopefully turn Partaj into an SPA, we need every one of our data
calls to go through it. It can start paying dividends even without
an SPA if we do the same data call more than once on the same view.

The cleanest way to implement this is to make individual hooks for
kinds of resources and relevant lists so we avoid repetition in
components.