🔥(frontend) remove the old useReferral hook
We previously built a useReferral hook to separate the data
fetching logic from our components. As we now intend to use react-query
for all data fetching, we build a new useReferral component that makes
use of it.

We can get rid of the old one, which is no longer used anywhere.