✨(frontend) add a button on draft answers to publish them
As we introduced a distinction on answer states (or kinds), between
the draft ones which can be revised and can only be seen by unit
members, and published ones which actually close the referral and
can be seen by the requester, we need a way to publish an answer
from a draft.

This is the new publish button, which reads as "send" to be more
easily understood by our users.
We also added a placeholder for the "revise" button, which we have
to implement next in the same spot.