✨(api) add reusable permission for objects related to referrals
We're working on new API routes and endpoints for objects related
to referrals (the first of which is ReferralAnswer). Some of the
permissions on ReferralAnswer can be expressed as a constraint on
properties of the related referral.

We expect this to happen for other objects later on. We decided to
create reusable permissions for this, as a clean way to express those
constraints.