♻️(frontend) show the answer form on the draft answer activity
As we built the referral detail view, we put the answer form at
the bottom of the list of all activities: it was just a UI component
that at that point was not associated with anything else on the
view.

Now that we actually create the referral answer draft beforehand
along with an activity, we would rather show it on the activity
itself, using the `ShowAnswerForm` context to know which answer
is currently being edited.