♻️(api) add the full author user object on answers
As the answer only contained only the author's id, we made some
awkward decisions to get around this without making extra requests.

For now, we'd rather just add the full user object. Later on, if
this is an issue we can revert it, this time making a separate
request for the user instead of iterating on the list of unit
members.