♻️(api) rework referral answer permissions
As we tried to use it for a retrieve call, we found out that our
ReferralAnswer permissions had an issue: as they relied on a
"referral" parameter in the request body, they could not work on
GET requests which supply no such parameter.

Conversely, POST requests cannot simply "get_object" like GET requests
can to determine the permissions.

As a result, we thought it better to make permissions more specific
to make them simpler and more expressive as we fixed this bug.