♻️(frontend) rework ReferralDetailAnswerForm into a live editor
In its second iteration (the first one that ran on the frontend),
<ReferralDetailAnswerForm /> was, as its name implied, just a form
that was powered by React.

Now, we refactored it completely to automatically save all changes
to the answer as it goes along, and the same goes for file uploads.

This was necessary to enable better interaction with draft answers
as well as revised versions of other answers.