🗃(models) change the attachment-answer relation to a ManyToMany
As we face the need to add answer revisions and published versions
of answers, we encountered a problem: how to manage attachments for
those answers we create duplicates of.

We chose to use a ManyToMany between answers and their attachments
so as to save uploading time and disk space with regards to the
actual files. This also still lets us keep track of which attachments
(and then files) are linked to an existing answer.