♻️(backend) update call sites for new answer-attachments relation
Some existing call sites, one in the API and all others in the
admin site, made direct use of the existing OneToMany relation
between referral answers and their attachments. We needed to update
them to use a related manager instead.

The call sites for the reverse relation are unaffected as those
already made use of the related manager.