✨(frontend) add a button to create an answer revision on answer display
We want to enable users to easily create a "clone" for an answer so
they can revise it as they please. As we already have an API endpoint
to create new instances of ReferralAnswer, we can use it here to
create the new revision, opening up the form as we succeed.