✨(api) add an endpoint to remove answer attachments
Answer attachment removal needs a specific endpoint beyond just
deleting the attachment objects: an attachment might be linked
with more than one answer, in which case deleting it is not an
acceptable way to solve this need.

We also would rather keep our endpoints simple, and not add too
much logic into our update endpoint, which is why we created a new
action endpoint specifically to unlink attachments.