✨(answer) add buttons to remove an attachment to the answer form
As our answer form is now a live editor, we actually create the
attachments on the backend as the user uploads them. This means
we need to add a mechanism to remove them in the live editor.

This is where our new <AnswerAttachmentsListEditor /> comes in,
showing the list of existing attachments and a button to remove
each of them from the current answer.