🐛(editor) fix an issue with forced update on the answer editor
The live editor exhibited a bug where it would lose the answer
content completely if, after a "Modify" interaction, the user forced
the update without having interacted with the editor at all.

This was due to a mismatch between the state machine context in
the answer form and the rich text field itself. We added an INIT
event that enables the field to immediately provide its starting
value to the parent component upon starting up. This way the
states are never out of sync.