✨(activity) update unit member activity permissions
When a unit member is also the referral linked user, they were
restricted to seeing only the activities a referral linked user can
see.

Instead, we'd rather they see all the activities just like a regular
unit member, as our use case for this is not that "it should not happen"
but rather that we can imagine a unit member creating the referral
themselves to track it in Partaj.