✨(api) add a search endpoint for users
To implement our validation request interface, we need users to
be able to pick any validator from Partaj (not just from their own
unit).

We intend to do this through an autocomplete list search endpoint.
To make it simple, we used django annotations and filters for now,
and may move to elasticsearch if we need more comprehensive query
or text transformations.