➕(frontend) add react-autosuggest to our dependencies
We just created a search list endpoint for users to enable unit
members to request a validation from any user by starting to type
in their name.

To build the frontend autocomplete UI that powers this feature, we'll
be using react-autosuggest as a battle-tested package that does all
the heavy lifting.