✨(frontend) add a user search UI to request validations on answers
As users can request validations for other users who are not members
of their unit, we need to enable them to "find" them, ie. get them
by name or email and ask them to validate their answers.

We held off on building tests for this component for now as we wait
for this interface to be accepted by our users before spending more
time on it.