✨(frontend) add a confirmation modal before sending an answer
Sending a referral answer is a sensitive action (as sending wrong data
may be harmful) and it is definitive (for now, as per the way we have
built Partaj).

We want to build some solemnity around it, to avoid user error and
prevent people from sending an answer by mistake.

To that end, we're adding a confirmation modal before users can send
an answer to the requester. We also use a btn-outline to open the modal
instead of a primary button so as not to draw too much attention to it.