🚧(api) hide the legacy test topic from API results
As we ramp out a new phase of tests for Partaj, we do not want
users to see the legacy test topic.

The ideal way to do this would be to add a boolean to topics,
determining if they accept new referral creation (and therefore
should appear in the list of topics).

As we need a very quick solution, we're just excluding the legacy
test topic by id for now.