🐛(activity) use validation request as foreign key on all validation activities
Some validation activities (validation requests) used ValidationRequest
objects as their generic foreign key, while others (approval, denial) used
ValidationResponse objects.

As we implemented the feature in Partaj's frontend, we found it simpler
to build using the request everywhere, especially as it includes the
response as a nested object, when there is one.