✨(frontend) implement referral answer validations
As we built our referral answer validation feature, we did not
implement the actual UI to perform the validation. We chose to add
a new form in a separate component to handle all the logic. We also
took this opportunity to improve the way we display existing
validations as well as their content.