🐛(referral) do not show error messages on empty referral form
Due to changes in the way our <RichTextField /> API is used, the
referral form was showing error messages for empty form fields
even when the user had not interacted with the form yet.

Update the referral form to use the new API so it does not consider
the field as dirty/cleaned when it has not been touched by the user.