🐛(referral) link published & draft answer as the draft is published
As we implemented referral answer publication through an action on
the referral model, we forgot to actually link together the draft
and the published version, which is necessary for things such as
validations (as those occur on drafts by definition).

Create a unit test ensuring this link is added, then make sure the
publish method creates it.