♻️(frontend) replace "context" with a lighter appData
The way we implemented our context (really application data as passed
to the frontend by the backend through the template) was as a property
picked up and passed down from the <Root /> component.

As Partaj grew, this became more and more troublesome as we had to
drill it everywhere in the app, even through multiple components that
had nothing to do with it. It was also unnecessary as it never changes
and so did not need to go through something dynamic like React props.

Instead, we can simply make one module that picks it up from the
DOM and makes it available by way of `import`.