✨(api) add a route to retrieve units
As part of our effort to move most of Partaj to a frontend app,
we need to be able to retrieve a unit in order to have its basic
information in the part of the app that works with units.

Open a retrieve endpoint that's only available for unit members.