✨(api) add an endpoint for processing tasks
We defined new entry points for a user to Partaj, based on what
the user is expected to accomplish. We're opening a new "tasks"
family of endpoints to get the content for these entry points.

One of these is to process the referrals they have been assigned
that are not yet under validation.