✨(api) add an endpoint for assignment tasks
We defined new entry points for a user to Partaj, based on what
the user is expected to accomplish. We're opening a new "tasks"
family of endpoints to get the content for these entry points.

One of these is to process the referrals units the user is an
owner of have received, that are not yet assigned to any user.