✨(api) add an endpoint for validation tasks
We defined new entry points for a user to Partaj, based on what
the user is expected to accomplish. We're opening a new "tasks"
family of endpoints to get the content for these entry points.

One of these is to perform validations on referral where they
have been requested to review answers.