✨(api) add more assignee data on referrals
The API was returning a list of IDs for the assignees on a given
referral. That was fine as long as we only every needed the assignees
for one referral at a time.

Now that we need to get them in list views of referral, we cannot make
an additional users request for each referral in the list. Instead, we
can use UserLite instances instead of ids to get the minimal
information we need.