🐛(frontend) fix a safari-only but where stretched links don't work
For referral lists, we need to display information as tables, which
creates an issue because rows cannot be links. What we can do is use
stretched links to extend the link over the whole row without breaking
the markup.

This does not work in Safari. We somewhat gave up for backend templates
as it was not practical to introduce JS to resolve the issue. With
the new referral lists in frontend React, it's much easier to add a JS
workaround.