✨(api) accept list requests for referrals for a given unit
As we move all views to the frontend, we now need a route to get
a unit's referrals, to show the same list we were showing in the
regular template in the unit view.

Enable the list endpoint for requests, but only with a unit params
and only for members of the given unit.