✨(frontend) add support for "no referrals yet" on <SentReferrals />
We previously only showed the link to the sent referrals view for users
who had already sent referrals. This meant that the view itself did not
need support for a case where there is no referral to display, as users
in this case had no "legitimate" way to get there.

This is more annoying to do in the frontend, and in our opinion it's not
necessarily the best choice. We'd rather keep a more stable UI and just
show a helpful message for users who have not yet created a referral, so
they can "look around".