♻️(backend) remove pass-through templates
As our frontend is ready to take over the app, we need to remove
all the pass-through templates that were used to load individual
view components via backend routing.

Then, we can add a unique App view and template that covers all
our frontend routes and loads a root component to that end.