✨(migrations) update the topic model to enable a hierarchy of topics
We want to add a lot of topics to Partaj. During this effort, we
realized we would need to structure them through hierarchical
relationships to keep them more manageable, and make them easier
to navigate for end users.

We're adding a parent field, so each topic can define what its own
parent is. We also took this opportunity to add a generic `is_active`
field to disable some topics without deleting them entirely.