✨(api) filter inactive & non-root topics from API calls
As topics are nested in the "children" property of their parent, we
do not need to return all children topics as part of the default
queryset (or they would be included twice).

We also took this opportunity to exclude inactive topics, which will
allow us to remove the "fix" that excluded expired topics.